Revert last 4 commits & move VanillaCommand to org.bukkit package
