+++ src/main/java/PluginBukkitBridge/entity/FakeCreeper.java
@@ -2,26 +2,25 @@ package PluginBukkitBridge.entity;
 
 import org.bukkit.entity.Creeper;
 
+import PluginBukkitBridge.MyPlugin;
 import PluginReference.MC_Entity;
 
 /**
  * Created by florian on 12.10.14.
  */
 public class FakeCreeper extends FakeCreature implements Creeper {
-    private boolean isPowered;
-    
     public FakeCreeper(MC_Entity argEnt) {
         super(argEnt);
-        this.isPowered = false;
     }
 
     @Override
     public boolean isPowered() {
-        return isPowered;
+        MyPlugin.fixme();
+        return false;
     }
 
     @Override
     public void setPowered(boolean b) {
-        this.isPowered = b;
+        MyPlugin.fixme();
     }
 }
