+++ src/main/java/PluginBukkitBridge/entity/FakeEnderCrystal.java
@@ -9,21 +9,19 @@ import org.bukkit.entity.EnderCrystal;
  * Created by florian on 12.10.14.
  */
 public class FakeEnderCrystal extends FakeEntity implements EnderCrystal {
-    private boolean isShowingBottom;
-    
     public FakeEnderCrystal(MC_Entity argEnt) {
         super(argEnt);
-        this.isShowingBottom = false;
     }
 
     @Override
     public boolean isShowingBottom() {
-        return this.isShowingBottom;
+        MyPlugin.fixme();
+        return true; // Minecraft Default
     }
 
     @Override
     public void setShowingBottom(boolean b) {
-        this.isShowingBottom = b;
+        MyPlugin.fixme();
     }
 
     @Override
