+++ src/main/java/PluginBukkitBridge/entity/FakeGuardian.java
@@ -1,27 +1,26 @@
 package PluginBukkitBridge.entity;
 
+import org.bukkit.entity.Guardian;
+
 import PluginBukkitBridge.MyPlugin;
 import PluginReference.MC_Entity;
-import org.bukkit.entity.Guardian;
 
 /**
  * Created by florian on 28.11.14.
  */
 public class FakeGuardian extends FakeCreature implements Guardian {
-    private boolean isElder;
-
 	public FakeGuardian(MC_Entity argEnt) {
 		super(argEnt);
-		this.isElder = false;
 	}
 
     @Override
     public boolean isElder() {
-        return isElder;
+        MyPlugin.fixme();
+        return false;
     }
 
     @Override
     public void setElder(boolean b) {
-        this.isElder = b;
+        MyPlugin.fixme();
     }
 }
