+++ src/main/java/org/bukkit/craftbukkit/help/SimpleHelpMap.java
@@ -1,18 +1,34 @@
 package org.bukkit.craftbukkit.help;
 
-import PluginBukkitBridge.FakeCraftServer;
-import PluginBukkitBridge.MyPlugin;
-import PluginBukkitBridge.commands.VanillaCommand;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.TreeMap;
+import java.util.TreeSet;
+
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.MultipleCommandAlias;
+import org.bukkit.command.PluginCommand;
+import org.bukkit.command.PluginIdentifiableCommand;
+import org.bukkit.command.defaults.BukkitCommand;
+import org.bukkit.command.defaults.VanillaCommand;
+import org.bukkit.help.GenericCommandHelpTopic;
+import org.bukkit.help.HelpMap;
+import org.bukkit.help.HelpTopic;
+import org.bukkit.help.HelpTopicComparator;
+import org.bukkit.help.HelpTopicFactory;
+import org.bukkit.help.IndexHelpTopic;
 
 import com.google.common.base.Predicate;
 import com.google.common.base.Predicates;
 import com.google.common.collect.Collections2;
-import org.bukkit.command.*;
-import org.bukkit.command.defaults.BukkitCommand;
-//import org.bukkit.command.defaults.VanillaCommand;
-import org.bukkit.help.*;
 
-import java.util.*;
+import PluginBukkitBridge.FakeCraftServer;
+import PluginBukkitBridge.MyPlugin;
 
 /**
  * Standard implementation of {@link HelpMap} for CraftBukkit servers.
